+++ CHANGELOG.md
@@ -2495,3 +2495,8 @@ Alternately, the _Daily Diary of my Programming Dreams_.
 
   1.  Added [fiddles/vue/fiddle-0012-VerticalSpacerJs](fiddles/vue/fiddle-0012-VerticalSpacerJs)
 
+
+### 1421-python-nice-gui-vega-embed ~ [Issue 1412](https://github.com/bradyhouse/house/issues/1412)
+
+* Added [fiddles/python/fiddle-0012-NiceGuiHW](fiddles/python/fiddle-0012-NiceGuiHW)
+